+++ examples/data-sources/fmc_intrusion_rules/data-source.tf
@@ -1,6 +1,0 @@
-data "fmc_intrusion_rules" "example" {
-  items = {
-    "my_intrusion_rules" = {
-    }
-  }
-}