+++ examples/resources/fmc_intrusion_rules/import.sh
@@ -1,3 +1,0 @@
-# <domain> is optional. If not provided, `Global` is used implicitly and resource's `domain` attribute is not set.
-terraform import fmc_intrusion_rules.example "<domain>,[<item1_name>,<item2_name>,...]"
-