+++ docs/guides/changelog.md
@@ -9,7 +9,7 @@ description: |-
 
 ## Unreleased
 
-- (Enhancement) Add support for Intrusion Rules and Intrusion Rule Group
+- (Enhancement) Add support for Intrusion Rule and Intrusion Rule Group
 - (Enhancement) Add support for Intrusion Policy Group Overrides
 
 ## 2.2.0